composer test entry point + autoload fix up
Use assignement in Autoload instead of array merge

When running jakub-onderka/php-parallel-lint, it chokes on:

    $GLOBALS['wgAutoloadClasses'] += array(

With:

    Error Output: PHP Notice:  Undefined index: wgAutoloadClasses
    PHP Fatal error:  Unsupported operand types

That is because the global is not set.

Rework the autoload to do explicit assignements, which would work even
when the global is not set.

Add a composer test entry point
Just run php lint for now.

Change-Id: I2a7b1bc7744ee49f8c48e5419077a3aabf3ef18e